Split AbsSeries.jl.
